Param speed to ints where needed for looping

+++ shaders/effects/synth/noise/definition.js
@@ -154,7 +154,7 @@ export default new Effect({
       }
     },
     speed: {
-      type: "float",
+      type: "int",
       default: 25,
       uniform: "speed",
       min: -100,
