+++ shaders/effects/synth/shape/definition.js
@@ -123,7 +123,7 @@ export default new Effect({
       }
     },
     speedA: {
-      type: "float",
+      type: "int",
       default: 50,
       uniform: "speedA",
       min: -100,
@@ -135,7 +135,7 @@ export default new Effect({
       }
     },
     speedB: {
-      type: "float",
+      type: "int",
       default: 50,
       uniform: "speedB",
       min: -100,
